### 2025-03-18 #7922 @editorialbot
[PRE REVIEW]: Tissue Data Explorer: a website template for presenting tissue sample research findings
**Submitting author:** @jamescarson3 (<a href="http://orcid.org/0000-0001-9009-5645">James P Carson</a>)
**Repository:** https://github.com/TACC/tissue-data-explorer
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
