### 2025-03-18 #7922 @jamescarson3
In addition to the auto-suggested most recent JOSS paper authors:
calebweninreb
magland

I suggest JOSS reviewers:
mikemcka
tomelse
andreped

And domain expert:
smcolby
