### 2025-04-29 804c89b
Add missing DOIs
