+++ paper/paper.bib
@@ -1,5 +1,6 @@
 @article{Jain:2023,
   	url = {https://doi.org/10.1038/s41556-023-01194-w},
+    doi = {10.1038/s41556-023-01194-w},
   	Author = {{Jain}, S. and {Pei}, L. and {Spraggins}, J.M. et al.},
     Journal = {Nat Cell Biol},
     volume = 25,
@@ -9,6 +10,7 @@
 
 @article{Börner:2022,
     url = {https://doi.org/10.1038/s42003-022-03644-x},
+    doi = {10.1038/s42003-022-03644-x},
     Author = {{Börner}, K. and {Bueckle}, A. and {Herr}, B.W. et al.},
     Title = {{Tissue registration and exploration user interfaces in support of a human reference atlas}},
     Journal = {Commun Biol},
@@ -17,26 +19,26 @@
 }
 
 @inproceedings{Ferlanti:2023,
-author = {Ferlanti, Erik and Allen, William J. and Lima, Ernesto A. B. F. and Wang, Yinzhi and Fonner, John M.},
-title = {Perspectives and Experiences Supporting Containers for Research Computing at the Texas Advanced Computing Center},
-year = {2023},
-isbn = {9798400707858},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-url = {https://doi.org/10.1145/3624062.3624587},
-doi = {10.1145/3624062.3624587},
-booktitle = {Proceedings of the SC '23 Workshops of the International Conference on High Performance Computing, Network, Storage, and Analysis},
-pages = {155–164},
-numpages = {10},
-keywords = {Apptainer, Containers, Docker, High performance computing, Singularity, System administration},
-location = {Denver, CO, USA},
-series = {SC-W '23}
+    author = {Ferlanti, Erik and Allen, William J. and Lima, Ernesto A. B. F. and Wang, Yinzhi and Fonner, John M.},
+    title = {Perspectives and Experiences Supporting Containers for Research Computing at the Texas Advanced Computing Center},
+    year = {2023},
+    isbn = {9798400707858},
+    publisher = {Association for Computing Machinery},
+    address = {New York, NY, USA},
+    url = {https://doi.org/10.1145/3624062.3624587},
+    doi = {10.1145/3624062.3624587},
+    booktitle = {Proceedings of the SC '23 Workshops of the International Conference on High Performance Computing, Network, Storage, and Analysis},
+    pages = {155–164},
+    numpages = {10},
+    keywords = {Apptainer, Containers, Docker, High performance computing, Singularity, System administration},
+    location = {Denver, CO, USA},
+    series = {SC-W '23}
 }
 
 @misc{Stack:2024,
-title = {Stack Overflow Developer Survey 2024},
-url={https://survey.stackoverflow.co/2024/},
-journal = {Stack Exchange Inc}
+    title = {Stack Overflow Developer Survey 2024},
+    url={https://survey.stackoverflow.co/2024/},
+    journal = {Stack Exchange Inc}
 }
 
 @article{Sorani:2012,
@@ -55,20 +57,21 @@ journal = {Stack Exchange Inc}
 }
 
 @inproceedings{Mariano:2020,
-author="Mariano, Diego
-and Ferreira, M{\'i}vian
-and Sousa, Bruno L.
-and Santos, Lucianna H.
-and de Melo-Minardi, Raquel C.",
-editor="Setubal, Jo{\~a}o C.
-and Silva, Waldeyr Mendes",
-title="A Brief History of Bioinformatics Told by Data Visualization",
-booktitle="Advances in Bioinformatics and Computational Biology",
-year="2020",
-publisher="Springer International Publishing",
[…]
