### 2025-04-29 #7922 @james-labyer
The suggested DOI for Docker Compose refers to a book about Docker Compose, not the software itself. The software itself does not have a DOI that I could find.
