### 2025-06-23 0bded17
update config portal upload validation to support listing PNG files as originals in metadata
