+++ config_portal/config_components/validate.py
@@ -273,89 +273,13 @@ def process_si_block_file(file: bytes, which_headers: str) -> tuple[bool, str]:
         return (header_check[0], header_check[1])
 
 
-def check_image_name(
-    filename: str, df: pd.DataFrame
-) -> tuple[int, str, pd.DataFrame, int, str]:
-    """This function checks that the provided image name can be associated with an entry
-    in the provided DataFrame. PNGs should be named like:
-
-    {Source file name}_C{Channel number, starting at zero}{sequence number, starting at zero and padded with zeroes to four digits}
-
-    Returns (number of matches, filename stem, df filtered to the relevant row, number of channels, error message)
-    """
-    ext = Path(filename).suffix
-    matches = 0
-    namestem = ""
-    if ext != ".png":
-        try:
-            matches = df["File"].eq(filename).sum()
-            namestem = Path(filename).stem
-            filename_reg = re.escape(filename)
-            new_df = df[df["File"].str.fullmatch(filename_reg)]
-            channel = None
-        except Exception as err:
-            return (
-                0,
-                False,
-                False,
-                False,
-                f"{err}",
-            )
-    else:
-        try:
-            nameparts = filename.split("_C")
-            if len(nameparts) < 2:
-                return (
-                    0,
-                    False,
-                    False,
-                    False,
-                    f"Filename {filename} must use the sequence _C to separate channel data",
-                )
-            namestem = "_C".join(nameparts[:-1])
-            namestem_reg = re.escape(namestem)
-            nr = namestem_reg + NAMEREG
-            matches = df["File"].str.contains(nr).sum()
-            new_df = df[df["File"].str.contains(nr)]
-        except Exception as err:
-            return (
-                0,
-                False,
-                False,
-                False,
-                f"{err}",
-            )
-        try:
-            channel = int(nameparts[-1][0])
-        except ValueError:
-            return (
-                0,
-                False,
-                False,
-                False,
-                "You must specify the channel for this image using the sequence _C as a separator",
-            )
-
-    if matches != 1:
-        return (
-            0,
-            False,
-            False,
-            False,
-            f"Filename {filename} must match exactly one filename in metadata",
-        )
-    else:
-        return matches, namestem, new_df, channel, ""
-
-
-def check_png_name_ending(filename: str) -> tuple[bool, str]:
+def check_png_name_ending(filename: str, nameparts: list) -> tuple[bool, str]:
     """Checks that the second half of a PNG follows the required naming convention:
 
     {Source file name}_C{Channel number, starting at zero}{sequence number, starting at zero and padded with zeroes to four digits}
 
     Returns (match test result, error message)
     """
-    nameparts = filename.split("_C")
     numreg = r"\d{5}\.png"
     nr = re.compile(numreg)
     m = nr.fullmatch(nameparts[-1])
@@ -366,6 +290,92 @@ def check_png_name_ending(filename: str) -> tuple[bool, str]:
             False,
             f"Filename {filename} must contain exactly five consecutive number characters between the _C sequence and the file extension",
         )
+
+
+def check_image_name(filename: str, df: pd.DataFrame) -> tuple[bool, str, str]:
+    """
+    Checks that the provided image name can be associated with an entry in the provided DataFrame.
+
+    If the image is a PNG that is not an original image, checks that the name matches the PNG naming convention.
+
+    Returns (name validity, namestem, error message)
+    """
+    ext = Path(filename).suffix
+    matches = 0
+    namestem = ""
+
+    try:
+        matches = df["File"].eq(filename).sum()
+        namestem = Path(filename).stem
+
+        if matches == 0 and ext == ".png":
+            try:
[…]
