### 2025-07-02 2b6f2c9
refactor image upload to handle source images with PNG format
