+++ config_portal/config_components/validate.py
@@ -283,7 +283,7 @@ def check_png_name_ending(filename: str, nameparts: list) -> tuple[bool, str]:
     numreg = r"\d{5}\.png"
     nr = re.compile(numreg)
     m = nr.fullmatch(nameparts[-1])
-    if m:
+    if m is not None:
         return True, ""
     else:
         return (
@@ -351,7 +351,6 @@ def get_image_info(
     if not valid_name[0]:
         return (False, "", pd.DataFrame({}), None, valid_name[1])
 
-    channel = None
     new_df = pd.DataFrame({})
 
     matches = df["File"].eq(filename).sum()
@@ -361,21 +360,10 @@ def get_image_info(
         new_df = df[df["File"].str.fullmatch(filename_reg)]
     else:
         # file validity has already been checked, so this must be a display image based on a file in metadata
-        nameparts = filename.split("_C")
         namestem_reg = re.escape(valid_name[1])
         nr = namestem_reg + NAMEREG
         new_df = df[df["File"].str.contains(nr)]
-        try:
-            channel = int(nameparts[-1][0])
-        except ValueError:
-            return (
-                False,
-                "",
-                pd.DataFrame({}),
-                None,
-                "You must specify the channel for this image using the sequence _C as a separator",
-            )
-    return (valid_name[0], valid_name[1], new_df, channel, "")
+    return (valid_name[0], valid_name[1], new_df, "")
 
 
 def process_sci_image(file: bytes, filename: str) -> tuple[bool, str, str]:
@@ -480,77 +468,66 @@ def make_thumbnail(imgs):
     return im
 
 
-def generate_thumbnails(idx, tdf):
-    # get image metadata
-    mdf = pd.read_csv(FD["si-block"]["si-files"]["publish"])
-    rd = {"Block": [], "Preview": [], "Name": [], "Link": []}
-    new_tr = pd.DataFrame(data=rd, dtype="string")
+def save_thumbnail(img, thumbnail_folder, thumbnail_loc):
+    if not Path.exists(thumbnail_folder):
+        Path.mkdir(thumbnail_folder, parents=True)
+    ImageOps.cover(img, FINAL_THUMBNAIL_SIZE).save(
+        thumbnail_loc,
+        optimize=True,
+    )
 
-    # get list of unique image sets from the thumbnail dataframe
-    parents = tdf["parent"].unique()
-    for i in range(parents.size):
-        # check how many channels are in the thumbnail dataframe for this parent, and if it matches metadata
-        igroup = tdf[tdf["parent"] == parents[i]]  # one row per channel
-        igroup = igroup[igroup["filename"].str.endswith(".png")]  # skip non-PNG files
-        # if no new PNG files, stop
-        if igroup.empty:
-            return
-        nr = re.escape(parents[i]) + NAMEREG
-        new_df = mdf[mdf["File"].str.contains(nr)]  # parent row in metadata
-        # how many channels should be in thumbnail
-        channels = new_df["Channels"].values[0]
-        imgs = []  # images to include in thumbnail
-        # compare number of rows in igroup to channels record in mdf
-        if igroup.shape[0] < channels:
-            # figure out which channel is missing
-            actual_channels = igroup["channel"].values
-            missing_channels = []
-            for k in range(channels):
-                if str(k) not in actual_channels:
-                    missing_channels.append(k)
-            # attempt to find images for any missing channels
-            for m in missing_channels:
-                p = Path(igroup["destination"].values[0])
-                pdir = p.parent
-                for file in pdir.iterdir():
-                    if file.is_file() and file.suffix == ".png":
-                        nameparts = file.name.split("_C")
-                        if int(nameparts[-1][0]) == m:
[…]
