### 2025-07-03 3741d96
Add ability to configure axis swap for 3d model files
