### 2025-07-07 29e90a4
Add support for tabs to 3d model page, support axis customization
