+++ app/pages/model3d.py
@@ -1,12 +1,23 @@
 import logging
+import json
 import dash_bootstrap_components as dbc
 import numpy as np
 import pandas as pd
 import plotly.graph_objects as go
-from dash import Input, Output, callback, dcc, html, register_page, ALL
+from dash import (
+    Input,
+    Output,
+    callback,
+    dcc,
+    html,
+    register_page,
+    MATCH,
+    callback_context,
+)
 from pywavefront import Wavefront
-
 from pages.constants import FILE_DESTINATION as FD
+from pages.ui import make_tabs
+
 
 register_page(__name__, path="/3d", title="3D Tissue Sample Model")
 
@@ -15,8 +26,44 @@ gunicorn_logger = logging.getLogger("gunicorn.error")
 app_logger.handlers = gunicorn_logger.handlers
 app_logger.setLevel(gunicorn_logger.level)
 
+# get block data
 blocks = pd.read_csv(FD["si-block"]["block-data"])
-traces = pd.read_csv(f"{FD["obj-files"]["summary"]}/obj-files.csv")
+
+
+def get_traces() -> tuple[list, list]:
+    """
+    Splits organ data into a separate Dataframe for each organ. This is necessary because when the traces are added to
+    separate graphs per organ, Dash will restart the indexing for the traces for each graph. Since the trace index is
+    the only way to identify the trace in the click data, the trace indexing needs to match in the graph and in the
+    Dataframe that will be used to populate the Block Data panel.
+
+    Returns organs, organ traces
+    """
+    traces = pd.read_csv(f"{FD["obj-files"]["summary"]}/obj-files.csv")
+
+    # split data by organ for later access
+    organs = list(traces["Organ"].unique())
+    organ_traces = []
+    for organ in organs:
+        # make a new dataframe and replace the index so we can find the traces based on their index within the graph
+        this_organ = traces.loc[traces["Organ"] == organ].copy()
+        this_organ.index = pd.Index([x for x in range(this_organ.shape[0])])
+        organ_traces.append(this_organ)
+    return organs, organ_traces
+
+
+organs, organ_traces = get_traces()
+
+
+def get_organ_desc(idx: int) -> str:
+    """Gets the organ description based on the organ index"""
+    row = blocks.loc[blocks["Organ ID"] == organ_traces[idx].loc[0, "Organ"]]
+    organ_desc = ""
+    if row.empty:
+        organ_desc = organ_traces[idx].loc[0, "Organ"]
+    else:
+        organ_desc = row["Organ Description"].unique()[0]
+    return organ_desc
 
 
 def check_null(value):
@@ -26,7 +73,8 @@ def check_null(value):
         return value
 
 
-def read_obj(file):
+def read_obj(file: str) -> tuple[np.array, np.array]:
+    """Reads an obj file at the given location and returns the vertices and faces of the object represented by the obj file."""
     organ = Wavefront(file, collect_faces=True)
     matrix_vertices = np.array(organ.vertices)
     faces = np.array(organ.mesh_list[0].faces)
@@ -37,8 +85,18 @@ def read_obj(file):
 
 
 def make_mesh_settings(
-    vertices, faces, name, color="cyan", opacity=1, x_map="x", y_map="y", z_map="z"
-):
+    vertices: np.array,
+    faces: np.array,
+    name: str,
+    color="cyan",
+    opacity=1,
+    x_map="x",
+    y_map="y",
+    z_map="z",
+) -> list[object]:
+    """
+    Populates a 3d mesh object for use with Plotly Dash graph objects
+    """
     coord_pos = {"x": 0, "y": 1, "z": 2}
 
[…]
