### 2025-07-07 f5f7e14
Update tests for multi tab 3d page
