+++ app/tests/test_3d.py
@@ -6,6 +6,8 @@ from dash import dcc, html
 import numpy as np
 import os
 import sys
+from dash._callback_context import context_value
+from dash._utils import AttributeDict
 
 if os.getcwd() not in sys.path:
     sys.path.append(os.getcwd())
@@ -84,36 +86,47 @@ def test_make_mesh_data():
 
 
 def test_make_mesh_fig():
-    fig1 = make_mesh_fig("S1")
+    fig1 = make_mesh_fig(0)
     assert len(fig1["data"]) == 9
     assert fig1["layout"]["height"] == 500
 
 
 def test_display_click_data():
-    click1 = [
-        {
-            "points": [
-                {
-                    "x": -392,
-                    "y": 1749.50244140625,
-                    "z": 746,
-                    "curveNumber": 0,
-                    "pointNumber": 52662,
-                    "i": 28916,
-                    "j": 28769,
-                    "k": 28768,
-                    "bbox": {
-                        "x0": 539.0667184592023,
-                        "x1": 539.0667184592023,
-                        "y0": 401.4491349814796,
-                        "y1": 401.4491349814796,
-                    },
+    def run_click_data(click):
+        context_value.set(
+            AttributeDict(
+                **{
+                    "triggered_inputs": [
+                        {"prop_id": '{"index":0,"type":"organ-graph"}.clickData'}
+                    ]
                 }
-            ]
-        }
-    ]
+            )
+        )
+        return display_click_data(click)
+
+    click1 = {
+        "points": [
+            {
+                "x": -392,
+                "y": 1749.50244140625,
+                "z": 746,
+                "curveNumber": 0,
+                "pointNumber": 52662,
+                "i": 28916,
+                "j": 28769,
+                "k": 28768,
+                "bbox": {
+                    "x0": 539.0667184592023,
+                    "x1": 539.0667184592023,
+                    "y0": 401.4491349814796,
+                    "y1": 401.4491349814796,
+                },
+            }
+        ]
+    }
+
     layout1 = json.loads(
-        json.dumps(display_click_data(click1), cls=plotly.utils.PlotlyJSONEncoder)
+        json.dumps(run_click_data(click1), cls=plotly.utils.PlotlyJSONEncoder)
     )
     testlayout1 = [
         dbc.CardHeader("Block Data"),
@@ -126,30 +139,29 @@ def test_display_click_data():
     test1 = json.loads(json.dumps(testlayout1, cls=plotly.utils.PlotlyJSONEncoder))
     assert layout1 == test1
 
-    click2 = [
-        {
-            "points": [
-                {
-                    "x": -10,
-                    "y": 0,
-                    "z": -5,
-                    "curveNumber": 4,
-                    "pointNumber": 0,
-                    "i": 0,
-                    "j": 1,
-                    "k": 3,
-                    "bbox": {
-                        "x0": 389.40599700379744,
-                        "x1": 389.40599700379744,
-                        "y0": 318.6925748500662,
-                        "y1": 318.6925748500662,
-                    },
-                }
-            ]
-        }
-    ]
+    click2 = {
+        "points": [
+            {
+                "x": -10,
+                "y": 0,
+                "z": -5,
+                "curveNumber": 4,
+                "pointNumber": 0,
+                "i": 0,
+                "j": 1,
+                "k": 3,
+                "bbox": {
+                    "x0": 389.40599700379744,
+                    "x1": 389.40599700379744,
+                    "y0": 318.6925748500662,
+                    "y1": 318.6925748500662,
+                },
+            }
+        ]
+    }
+
     layout2 = json.loads(
[…]
