### 2025-07-18 038c8db
fix thumbnail generation bug
