+++ config_portal/config_components/validate.py
@@ -12,7 +12,6 @@ from PIL import Image, ImageDraw, ImageOps
 import logging
 import base64
 from collections.abc import Callable
-
 from pages import constants
 
 MAX_TITLE_LENGTH = 2048
@@ -356,7 +355,7 @@ def get_image_info(
     """
     Validates the provided image, and if valid, gets information associated with the image.
 
-    Returns (name validity, filename stem, df filtered to the relevant row, image channel, error message)
+    Returns (name validity, filename stem, df filtered to the relevant row, error message)
     """
     valid_name = check_image_name(filename, df)
     if not valid_name[0]:
@@ -436,15 +435,56 @@ def update_title(value: str) -> tuple[str, str, str]:
         return ("Title not updated", "Please provide a name.", "failure")
 
 
+def query_thumbnails_for_changed_filenames(
+    blocks_df: pd.DataFrame, tn_df: pd.DataFrame
+) -> pd.Series:
+    """
+    Returns a list of filenames
+    """
+    # make a df shaped like thumbnails df based on blocks df
+    new_td_df = pd.DataFrame(data={})
+    new_td_df["Block"] = blocks_df["Tissue Block"]
+    new_td_df["Name"] = blocks_df["File"]
+    new_td_df["Preview"] = ""
+    new_td_df["Link"] = (
+        f"/scientific-images/{blocks_df["Tissue Block"]}/{blocks_df["Image Set"]}"
+    )
+
+    # combine old and new dfs
+    combined_df = pd.concat([new_td_df, tn_df], ignore_index=True)
+
+    # Filter down to just files that were in the blocks df
+    just_new = combined_df.loc[combined_df["Name"].isin(blocks_df["File"])]
+
+    # drop rows that have not changed in the relevant columns
+    changed_entries = just_new.drop_duplicates(
+        keep=False, inplace=False, ignore_index=True, subset=["Block", "Name", "Link"]
+    )
+
+    return changed_entries["Name"]
+
+
 def publish_si_block() -> tuple[str, str, str]:
-    """Publish scientific images metadata.
+    """Publish scientific images metadata. Triggers re-generation of thumbnails if the image definition has changed.
+
     Returns (title of update toast, description of update, success status)"""
     try:
+        new_si_df = pd.read_csv(FD["si-block"]["si-files"]["depot"])
+        old_tn_df = pd.read_csv(FD["thumbnails"]["catalog"])
+
+        # find out which entries in the new blocks file will need to be updated in the thumbnails catalog
+        changed_entries = query_thumbnails_for_changed_filenames(new_si_df, old_tn_df)
+        files = list(changed_entries)
+
+        generate_thumbnails(files, new_si_df, False)
+
+        # put new scientific images metadata in publish location
         for key in FD["si-block"].keys():
             p = Path(FD["si-block"][key]["depot"])
             t = Path(FD["si-block"][key]["publish"])
             if p.exists():
                 shutil.move(p, t)
+
     except FileNotFoundError:
         return ("Metadata not updated", "File not found", "failure")
     return (
@@ -488,51 +528,116 @@ def save_thumbnail(img, thumbnail_folder, thumbnail_loc):
     )
 
 
-def generate_thumbnails(tdf: pd.DataFrame):
+def generate_thumbnail(
+    dest: str, img_set: str, thumbnail_loc: str
+) -> tuple[bool, bool, str]:
     """
-    Generate thumbnails for images that have been updated.
+    Generate a thumbnail for a scientific image and save the image.
+
+    Returns (success, error message)
     """
-
-    # tdf = {"name": [], "namestem": [], "block": [], "imageset": [], "destination": []}
[…]
