### 2025-07-23 3756f9d
refactor 3d model page for faster loading
