+++ app/pages/model3d.py
@@ -7,6 +7,7 @@ import plotly.graph_objects as go
 from dash import (
     Input,
     Output,
+    State,
     callback,
     dcc,
     html,
@@ -16,7 +17,7 @@ from dash import (
 )
 from pywavefront import Wavefront
 from pages.constants import FILE_DESTINATION as FD
-from pages.ui import make_tabs
+from pages.ui import make_tabs, make_loader
 
 
 register_page(__name__, path="/3d", title="3D Tissue Sample Model")
@@ -30,7 +31,28 @@ app_logger.setLevel(gunicorn_logger.level)
 blocks = pd.read_csv(FD["si-block"]["block-data"])
 
 
-def get_traces() -> tuple[list, list]:
+def get_organs() -> list:
+    # Get organ ID's
+    traces = pd.read_csv(f"{FD["obj-files"]["summary"]}/obj-files.csv")
+    organs = list(traces["Organ"].unique())
+    organ_descs = []
+    organ_traces = []
+
+    # Get organ descriptions and trace name
+    for i in range(len(organs)):
+        desc_block = blocks.loc[blocks["Organ ID"] == organs[i]]
+        if desc_block.empty:
+            desc = organs[i]
+        else:
+            desc = desc_block["Organ Description"].unique()[0]
+        organ_descs.append(desc)
+        this_organ = traces.loc[traces["Organ"] == organs[i]].copy()
+        organ_traces.append(this_organ["Name"].to_list())
+
+    return organ_descs, organ_traces
+
+
+def get_trace(idx) -> tuple[list, list]:
     """
     Splits organ data into a separate Dataframe for each organ. This is necessary because when the traces are added to
     separate graphs per organ, Dash will restart the indexing for the traces for each graph. Since the trace index is
@@ -43,27 +65,9 @@ def get_traces() -> tuple[list, list]:
 
     # split data by organ for later access
     organs = list(traces["Organ"].unique())
-    organ_traces = []
-    for organ in organs:
-        # make a new dataframe and replace the index so we can find the traces based on their index within the graph
-        this_organ = traces.loc[traces["Organ"] == organ].copy()
-        this_organ.index = pd.Index([x for x in range(this_organ.shape[0])])
-        organ_traces.append(this_organ)
-    return organs, organ_traces
-
-
-organs, organ_traces = get_traces()
-
-
-def get_organ_desc(idx: int) -> str:
-    """Gets the organ description based on the organ index"""
-    row = blocks.loc[blocks["Organ ID"] == organ_traces[idx].loc[0, "Organ"]]
-    organ_desc = ""
-    if row.empty:
-        organ_desc = organ_traces[idx].loc[0, "Organ"]
-    else:
-        organ_desc = row["Organ Description"].unique()[0]
-    return organ_desc
+    this_organ = traces.loc[traces["Organ"] == organs[idx]].copy()
+    this_organ.index = pd.Index([x for x in range(this_organ.shape[0])])
+    return this_organ
 
 
 def check_null(value):
@@ -157,21 +161,22 @@ def make_mesh_data(
 
 def make_mesh_fig(idx=0) -> go.Figure:
     """Plots the objects described in an obj in a Plotly Dash figure"""
-    if organ_traces[idx].shape[0] == 0:
+    organ_trace = get_trace(idx)
+    if organ_trace.shape[0] == 0:
         return
     start = True
-    for i in organ_traces[idx].index:
+    for i in organ_trace.index:
         try:
-            file_loc = f"{FD["obj-files"]["volumes"]}/{organ_traces[idx].at[i, "File"]}"
+            file_loc = f"{FD["obj-files"]["volumes"]}/{organ_trace.at[i, "File"]}"
             if start:
[…]
