### 2025-07-28 1136db4
Add caching to 3d model page
