+++ pyproject.toml
@@ -20,6 +20,7 @@ pywavefront = "~1.3.3"
 pillow = "~11.0.0"
 ruff = "~0.8.2"
 nh3 = "^0.2.20"
+flask-caching = "^2.3.1"
 
 [tool.poetry.group.dev]
 optional = true
