### 2025-07-29 99001a8
add support for reports grouped by organ
