+++ app/tests/test_oc-links.py
@@ -5,10 +5,10 @@ import pandas as pd
 if os.getcwd() not in sys.path:
     sys.path.append(os.getcwd())
 import app
-from pages.ocpage import s_imgs
 from pages.constants import FILE_DESTINATION as FD
 
 thumbnails = pd.read_csv(FD["thumbnails"]["catalog"])
+s_imgs = pd.read_csv(FD["si-block"]["si-files"])
 
 
 def test_oc_page_links():
