### 2025-08-05 b566ba7
fix pandas dataframe concatenation warning
