+++ config_portal/config_components/validate.py
@@ -12,7 +12,7 @@ from PIL import Image, ImageDraw, ImageOps
 import logging
 import base64
 from collections.abc import Callable
-from pages import constants
+from pages.constants import FILE_DESTINATION as FD
 
 MAX_TITLE_LENGTH = 2048
 MAX_FILENAME_LENGTH = 255
@@ -53,7 +53,6 @@ VALID_EXTS = {
     ],
 }
 
-
 VALID_MIMES = {
     "excel": [
         "application/vnd.ms-excel",
@@ -73,6 +72,7 @@ VALID_MIMES = {
     ],
     "image": ["image/tiff", "image/png", "video/x-msvideo", "image/jpeg"],
 }
+
 REQUIRED_HEADERS = {
     "si-block": {
         "block-data": [
@@ -145,8 +145,6 @@ REQUIRED_HEADERS = {
     },
     "reports": {"reports": ["Name", "Organ ID", "Organ Description", "Link"]},
 }
-
-FD = constants.FILE_DESTINATION
 
 app_logger = logging.getLogger(__name__)
 gunicorn_logger = logging.getLogger("gunicorn.error")
@@ -528,6 +526,29 @@ def save_thumbnail(img, thumbnail_folder, thumbnail_loc):
     )
 
 
+def collect_pngs_for_thumbnail(dest: str) -> list:
+    """
+    Checks the destination folder for PNG images to use for generating a thumbnail and returns selected images.
+    """
+    p = Path(dest)
+    display_imgs = []
+    channels = []
+
+    # collect the images with PNG naming scheme
+    if Path.exists(p):
+        for file in p.iterdir():
+            if file.is_file() and file.suffix == ".png":
+                nameparts = file.name.split("_C")
+                is_display_img = check_png_name_ending(file.name, nameparts)
+                if is_display_img[0]:
+                    # check if this channel is already represented in the list
+                    if nameparts[-1][0] not in channels:
+                        # add it to list if not
+                        channels.append(nameparts[-1][0])
+                        display_imgs.append(Image.open(file))
+    return display_imgs
+
+
 def generate_thumbnail(
     dest: str, img_set: str, thumbnail_loc: str
 ) -> tuple[bool, bool, str]:
@@ -537,22 +558,7 @@ def generate_thumbnail(
     Returns (success, error message)
     """
     try:
-        p = Path(dest)
-        display_imgs = []
-        channels = []
-
-        # collect the images with PNG naming scheme
-        if Path.exists(p):
-            for file in p.iterdir():
-                if file.is_file() and file.suffix == ".png":
-                    nameparts = file.name.split("_C")
-                    is_display_img = check_png_name_ending(file.name, nameparts)
-                    if is_display_img[0]:
-                        # check if this channel is already represented in the list
-                        if nameparts[-1][0] not in channels:
-                            # add it to list if not
-                            channels.append(nameparts[-1][0])
-                            display_imgs.append(Image.open(file))
+        display_imgs = collect_pngs_for_thumbnail(dest)
 
         # Delete old thumbnail if exists. This is necessary because the image set could be changed to point to a different image.
         thumbnail_folder = Path(f"{FD['thumbnails']['publish']}/{img_set}")
@@ -891,20 +897,23 @@ def make_cubes_df(
             z_transform - 0.001,
         ],
     }
-    cubes_df = pd.DataFrame().reindex(columns=points_df.columns)
+    cubes_df = pd.DataFrame(columns=points_df.columns).astype(dtype=points_df.dtypes)
 
[…]
