### 2025-08-05 95c0c13
refactor for readability
