### 2025-08-13 cf3f7a9
add ability to upload images and view them on volumetric map
