+++ app/tests/test_spatialmap.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import numpy as np
 import pytest
 import os
 import sys
@@ -12,14 +13,20 @@ from pages.spatialmap import (
     make_layers,
     load_data,
     update_fig,
+    make_image_layers,
+    get_colorscale,
+    get_all_images,
 )
+from pages.ui import make_image_layer_fig
 
 D_PROTEIN = "CYB5A"
 D_SCHEME = "jet"
 D_OPACITY = 0.4
 D_LAYER = "All"
 
-page_info, defaults, layers, cat_opts, value_info, axes, downloads = load_data("S1-12")
+page_info, defaults, layers, image_layers, cat_opts, value_info, axes, downloads = (
+    load_data("S1-12")
+)
 
 
 def test_make_defaults():
@@ -31,6 +38,7 @@ def test_make_defaults():
     expected = {
         "d_scheme": "haline",
         "d_layer": "All",
+        "i_layer": "All",
         "d_category": "All",
         "d_value": "CYB5A",
     }
@@ -60,9 +68,30 @@ def test_make_layers():
     assert make_layers(z_axis) == expected_layers
 
 
+def test_make_image_layers():
+    # make_image_layers(df: pd.DataFrame, block)
+    data = {"Block": ["A", "A", "B"]}
+    df = pd.DataFrame(data=data)
+    a_expected_layers = ["All", "Image 1", "Image 2"]
+    b_expected_layers = ["All", "Image 1"]
+
+    assert make_image_layers(df, "A") == a_expected_layers
+    assert make_image_layers(df, "B") == b_expected_layers
+
+
+def test_get_colorscale():
+    expected_scale = [
+        [np.float64(0.0), "rgb(255,0,0)"],
+        [np.float64(0.67), "rgb(0,0,255)"],
+        [np.float64(0.99), "rgb(0,255,0)"],
+        [np.float64(1.0), "rgb(255,255,255)"],
+    ]
+    assert get_colorscale("channel_split") == expected_scale
+
+
 def test_load_data():
-    page_info, defaults, layers, cat_opts, value_info, axes, downloads = load_data(
-        "S1-12"
+    page_info, defaults, layers, image_layers, cat_opts, value_info, axes, downloads = (
+        load_data("S1-12")
     )
     expected_page_info = {
         "Block": "S1-12",
@@ -73,12 +102,15 @@ def test_load_data():
     expected_defaults = {
         "d_scheme": "haline",
         "d_layer": "All",
+        "i_layer": "All",
         "d_category": "All",
         "d_value": "CYB5A",
     }
     assert defaults == expected_defaults
     expected_layers = ["All", "Layer 1", "Layer 2", "Layer 3", "Layer 4"]
     assert layers == expected_layers
+    expected_image_layers = ["All", "Image 1", "Image 2", "Image 3", "Image 4"]
+    assert image_layers == expected_image_layers
     expected_cat_opts = {
         "Category": "Gland",
         "Label (Only True)": "Pixels with gland tissue",
@@ -116,104 +148,206 @@ def test_load_data():
 
 
 def test_uf_cube_opacity():
-    # test that fig1's opacity value is updated as expected
-    fig1 = update_fig(
-        "cube-tab",
-        D_SCHEME,
-        D_PROTEIN,
-        0.5,
-        0.4,
-        "All",
-        "All",
-        category_data=cat_opts,
-        value_range_dict=value_info,
-        axes=axes,
-        block="S1-12",
-    )
-    fig2 = update_fig(
-        "cube-tab",
-        D_SCHEME,
-        D_PROTEIN,
-        0,
-        0.4,
-        "All",
-        "All",
-        category_data=cat_opts,
-        value_range_dict=value_info,
-        axes=axes,
-        block="S1-12",
-    )
-
[…]
