+++ app/pages/spatialmap.py
@@ -13,9 +13,11 @@ from dash import (
 import pandas as pd
 import numpy as np
 from pathlib import Path
+from PIL import Image
 from pages.constants import FILE_DESTINATION as FD
 from components import alerts
 import pages.ui as ui
+import sys
 
 app_logger = logging.getLogger(__name__)
 gunicorn_logger = logging.getLogger("gunicorn.error")
@@ -38,10 +40,13 @@ register_page(
 
 
 # Initial data retrieval tasks
-
-
 def make_defaults(ranges_df: pd.DataFrame) -> dict:
-    defaults = {"d_scheme": "haline", "d_layer": "All", "d_category": "All"}
+    defaults = {
+        "d_scheme": "haline",
+        "d_layer": "All",
+        "i_layer": "All",
+        "d_category": "All",
+    }
     d_val = ranges_df.columns[np.nonzero(ranges_df.loc["Default"])].values[0]
     defaults["d_value"] = d_val
     return defaults
@@ -68,6 +73,48 @@ def make_layers(z_axis: list) -> list:
     return layers
 
 
+def make_image_layers(df: pd.DataFrame, block) -> dict:
+    # get entries for this block from df
+    this_block = df.loc[df["Block"] == block]
+    # get count of items in this block
+    num_entries = this_block["Block"].count()
+    # make layers list
+    layers = ["All"]
+    layers.extend([f"Image {i + 1}" for i in range(num_entries)])
+    return layers
+
+
+def get_all_images(block):
+    img_info = ui.get_image_info(block)
+    imgs = []
+    files = list(img_info["Name"])
+    for file in files:
+        fpath = Path(file)
+        fstem = fpath.stem
+        p = Path(f"{FD["image-layer"]}/{block}/layers/{fstem}.txt")
+        img = np.loadtxt(p, delimiter="\t")
+        imgs.append(img)
+    return img_info, imgs
+
+
+def get_colorscale(scale: str) -> list:
+    cs_df = pd.read_csv(f"{FD["image-layer"]}/colorscales.csv")
+    # filter down to just that colorscale
+    this_scale = cs_df.loc[cs_df["Scale Name"] == scale]
+
+    if this_scale.size > 0:
+        entries = this_scale["Entry"].to_list()
+        formatted_scale = []
+        for entry in entries:
+            scale_row = this_scale.loc[this_scale["Entry"] == entry]
+            scale_marker = scale_row["Scale Marker"].values[0]
+            scale_color = f"rgb({scale_row["R"].values[0]},{scale_row["G"].values[0]},{scale_row["B"].values[0]})"
+            formatted_scale.append([scale_marker, scale_color])
+        return formatted_scale
+    else:
+        return None
+
+
 def load_data(block: str) -> tuple[dict, dict, list, list, list, dict]:
     # page info dict, defaults dict, layers, category options, values, axes
     dir = f"{FD['volumetric-map']}/{block}"
@@ -76,6 +123,7 @@ def load_data(block: str) -> tuple[dict, dict, list, list, list, dict]:
     category_labels = pd.read_csv(f"{dir}/category_labels.csv")
     vol_measurements = pd.read_csv(f"{dir}/vol_measurements.csv")
     downloads = pd.read_csv(f"{FD['volumetric-map']}/downloads.csv")
+    image_layers_df = pd.read_csv(f"{FD["image-layer"]}/images.csv")
 
     # Get title and desc
     page_info = meta.iloc[0].to_dict()
@@ -89,20 +137,39 @@ def load_data(block: str) -> tuple[dict, dict, list, list, list, dict]:
     # Make layers list
     layers = make_layers(axes["Z"])
 
+    # Make image layers dict
+    image_layers = make_image_layers(image_layers_df, block)
+
[…]
