### 2025-08-14 af57ad0
Add test for convert_img_to_greyscale_array
