### 2025-08-15 d96ce99
fix reports page access error
