### 2025-08-15 d279d2b
update cleanup instructions to just clean up tde related items
