### 2025-08-26 1f9251e
refactor to run Docker container with a nonroot user
