### 2025-08-26 62861be
Add instructions for contributing to the project
