+++ CONTRIBUTING.md
@@ -1,0 +1,58 @@
+
+# How to Contribute
+
+Thank you for taking the time to contribute to Tissue Data Explorer!
+
+A project overview and setup instructions can be found in the [README](./README.md).
+
+## Contact info
+Please contact [James Labyer](james.labyer@austin.utexas.edu) with questions about contributing to the project.
+
+## Issues
+We use [GitHub issues](https://github.com/TACC/tissue-data-explorer/issues) in this repository to track both bug reports and feature requests.
+
+Before creating a new issue, please search for similar issues that already exist. If you find one, please add a comment to the existing issue instead of creating a new issue.
+
+If you have a fix ready for an issue you are reporting, please open a pull request with the fix, and include a link to it in the issue.
+
+### Bugs
+When reporting a bug, please include as much information about the problem as possible, including:
+- A clear and thorough description of the observed behavior
+- The behavior that you expect to see instead of the observed behavior
+- Steps to reproduce the problem
+- Conditions that cause the problem to disappear
+- Information about how you are running the software. Do you have both the display site and configuration portal set up? Which Docker Compose file are you using to run the container where you are seeing the behavior?
+
+### Enhancement Requests
+When requesting a new feature, please include:
+- A description of how the feature would be used
+- An explanation for how the feature would affect your workflows
+- Your willingness to contribute to the enhancement
+
+## Code Standards
+When contributing code to the project, please follow these standards:
+- Create a new branch and make your changes there, make sure that the automated tests pass, then open a PR to merge your branch into main.
+- Your code will be reviewed before it is merged. Please allow some time for the review.
+- This project uses ruff as a linter. Please see [pyproject.toml](./pyproject.toml) for rule exceptions
+
+## Community and Behavioral Expectations
+Adapted from the [Contributor Covenant Code of Conduct](https://www.contributor-covenant.org/version/2/1/code_of_conduct/)
+
+### Our Pledge
+We as members, contributors, and leaders pledge to make participation in our
+community a harassment-free experience for everyone. We pledge to act and interact 
+in ways that contribute to an open and welcoming community.
+
+### Our Standards
+Examples of behavior that contributes to a positive environment for our
+community include:
+* Demonstrating empathy and kindness toward other people
+* Being respectful of differing opinions, viewpoints, and experiences
+* Giving and gracefully accepting constructive feedback
+* Accepting responsibility and apologizing to those affected by our mistakes,
+and learning from the experience
+* Focusing on what is best not just for us as individuals, but for the overall
+community
+
+Unacceptable behavior includes conduct which could reasonably be considered 
+inappropriate in a professional setting.
