### 2025-08-28 f028898
add scripts to streamline build process
