+++ run_tests.sh
@@ -1,20 +1,11 @@
 #!/bin/bash
 
-USERID=$(id -u)
-GROUPID=$(id -g)
-
-docker volume create config-data-test
-docker run -d --name temp -v config-data-test:/data busybox sleep infinity
-docker cp ./start/. temp:/data/
-docker stop temp
-USERID=$USERID GROUPID=$GROUPID docker compose -f docker-compose-test.yaml build &> build.log
-USERID=$USERID GROUPID=$GROUPID docker compose -f docker-compose-test.yaml up --detach
-docker exec tissue-data-explorer-display-1 bash -c "python -m pytest"
-docker stop tissue-data-explorer-display-1
-docker rm tissue-data-explorer-display-1
-docker exec tissue-data-explorer-config-1 bash -c "python -m pytest"
-docker stop tissue-data-explorer-config-1
-docker rm tissue-data-explorer-config-1
-docker rm temp
+/bin/bash ./build.sh -e test -v config-data-test -n
+docker exec tde-dev-display-1 bash -c "python -m pytest"
+docker stop tde-dev-display-1
+docker rm tde-dev-display-1
+docker exec tde-dev-config-1 bash -c "python -m pytest"
+docker stop tde-dev-config-1
+docker rm tde-dev-config-1
 docker volume rm config-data-test
 
