+++ docker-compose-test.yaml
@@ -1,20 +1,0 @@
-# This compose file is intended for running tests in a development environment
-# using the run_tests.sh script, which creates a volume and populates it with
-# test data so the tests always run on the same data.
-
-services:
-  display:
-    extends:
-      file: docker-compose-dev.yaml
-      service: display-app
-    volumes:
-    - config-data-test:/home/nonroot/app/assets/config
-  config:
-    extends:
-      file: docker-compose-dev.yaml
-      service: config-app
-    volumes:
-    - config-data-test:/config
-volumes:
-  config-data-test:
-    external: true
