### 2025-08-28 9316aa7
Edits to address JOSS feedback
