+++ paper/paper.md
@@ -45,17 +45,17 @@ bibliography: paper.bib
 
 # Summary
 
-The Tissue Data Explorer is a Python code repository that can be used to stand up a website for displaying 2D and 3D biomolecular datasets that have been collected from tissue samples taken from organs [@Labyer:2025]. The Tissue Data Explorer includes interfaces for loading and displaying image stacks, spatial maps of measurements taken from different areas within the tissue sample, 3D models of the tissue sample, and links to custom reports. Life sciences researchers can stand up an instance of the website as is, or they can fork the repository and modify the interfaces to accommodate additional data types. To enable easier customization, the website is written using the Plotly Dash visualization framework [@Parmer:2024]. Dash allows users to build visuals using Python, which is likely to be easier for many biologists because Python knowledge is common among life sciences researchers [@Mariano:2020] [@Boudreau:2021]. This is in contrast with other common web languages like JavaScript, which is popular among software developers [@Stack:2024] but has not achieved the same level of adoption with life sciences researchers [@Sorani:2012].
+The Tissue Data Explorer is a Python code repository that can be used to deploy a website for displaying 2D and 3D biomolecular datasets that have been collected from tissue samples taken from organs [@Labyer:2025]. The Tissue Data Explorer includes interfaces for loading and displaying image stacks, spatial maps of measurements taken from different areas within the tissue sample, 3D models of the tissue sample, and links to custom reports. Life sciences researchers can deploy an instance of the website as is, or they can fork the repository and modify the interfaces to accommodate additional data types. To enable easier customization, the website is written using the Plotly Dash visualization framework [@Parmer:2024]. Dash allows users to build visuals using Python, which is likely to be easier for many biologists because Python knowledge is common among life sciences researchers [@Mariano:2020] [@Boudreau:2021]. This is in contrast with other common web languages like JavaScript, which is popular among software developers [@Stack:2024] but has not achieved the same level of adoption with life sciences researchers [@Sorani:2012].
 
 # Statement of Need
 
-This software emerged from a project which aimed to provide interactive visualization of data collected by a Human BioMolecular Atlas Program  (HuBMAP) [@Jain:2023] pancreas Tissue Mapping Center to facilitate exploration of the datasets via web-browser. We realized that other research teams have collected or will collect similar data sets on other types of tissue, and may also want to provide interactive visualizations of their own datasets. While there do exist large, centralized websites that host biology datasets and provide users with interactive data analysis capabilities [@Börner:2022], these other services do not provide researchers wide control over how the data is displayed. Our software gives researchers display interfaces out of the box, but also provides a model for how researchers may create other visualizations as they see fit for their own needs. While we developed this software for a future publication of pancreas data, we present here the Tissue Data Explorer along with provided synthetic sample data. 
+This software emerged from a project that aimed to provide interactive visualization of data collected by a Human BioMolecular Atlas Program  (HuBMAP) [@Jain:2023] pancreas Tissue Mapping Center to facilitate exploration of the datasets via web-browser. We realized that other research teams have collected or will collect similar datasets on other types of tissue, and may also want to provide interactive visualizations of their own datasets. While there do exist large, centralized websites that host biology datasets and provide users with interactive data analysis capabilities [@Börner:2022], these other services do not provide researchers wide control over how the data is displayed. Our software gives researchers display interfaces out of the box, but also provides a model for how researchers may create other visualizations as they see fit for their own needs. While we developed this software for a future publication of pancreas data, we present here the Tissue Data Explorer along with provided synthetic sample data. 
 
 # Main Features
 
 The software package consists of:
 
-- a configuration app, which requires authentication and allows deployers to upload the data that will display on the website
[…]
