### 2025-09-29 3e00305
Update docker restart policy to always so that services will restart on server reboot
