### 2025-10-14 TACC/tissue-data-explorer#18 @magland
installation issue
Tried to install using the installation instructions, but when I ran ./build.sh I got

```
$ ./build.sh
unknown shorthand flag: 'f' in -f

Usage:  docker [OPTIONS] COMMAND [ARG...]

Run 'docker --help' for more information
```

I am on fedora linux with

Docker version 28.4.0, build 1.fc42
