### 2025-10-20 TACC/tissue-data-explorer#18 @james-labyer
Hi there! The error message suggests an issue with your Docker compose installation. Could you verify that Docker compose is working for you by running `docker compose version`? You may need to install the Docker compose plugin for Fedora. Here are the [instructions](https://docs.docker.com/compose/install/linux/) for installing the plugin. This [GitHub issue](https://github.com/docker/compose/issues/8986) on a similar error message suggests some other solutions as well. Let me know if you are still getting this error message once you have verified your Docker compose installation.
