### 2025-10-21 TACC/tissue-data-explorer#19 @magland
error getting started running the server
I encounter the following error when following the instructions for getting started

```bash
$ ./build.sh -n
config-data-dev
237a3776f1dc703639f01612481022f189a608720f38783dba5f175d2a4a05de
Successfully copied 45.2MB to temp:/data/
temp
temp
[+] Building 1.8s (8/25)
 => [internal] load local bake definitions                                                                                                                                                                                                                                                                                                    0.0s
 => => reading from stdin 1.21kB                                                                                                                                                                                                                                                                                                              0.0s
 => [display internal] load build definition from Dockerfile-all                                                                                                                                                                                                                                                                              0.1s
 => => transferring dockerfile: 3.76kB                                                                                                                                                                                                                                                                                                        0.0s
 => [display internal] load metadata for docker.io/library/python:3.13-bookworm                                                                                                                                                                                                                                                               0.5s
 => [display internal] load .dockerignore                                                                                                                                                                                                                                                                                                     0.1s
 => => transferring context: 138B                                                                                                                                                                                                                                                                                                             0.0s
 => [config internal] load build context                                                                                                                                                                                                                                                                                                      0.1s
 => => transferring context: 186B                                                                                                                                                                                                                                                                                                             0.0s
 => CACHED [config poetry 1/5] FROM docker.io/library/python:3.13-bookworm@sha256:244755764f644befee8fe3114a13009ff94d76cf2f399033cdf87974df7fbcba                                                                                                                                                                                            0.0s
 => CANCELED [config poetry 2/5] RUN apt-get update && apt-get upgrade -y                                                                                                                                                                                                                                                                     0.8s
 => ERROR [config dev-config 2/8] RUN adduser --uid 1000 --gid 1000 --disabled-password --gecos "" nonroot                                                                                                                                                                                                                                    0.6s
------
 > [config dev-config 2/8] RUN adduser --uid 1000 --gid 1000 --disabled-password --gecos "" nonroot:
0.516 adduser: The GID 1000 does not exist.
------
Dockerfile-all:14

--------------------

  12 |     ARG USERID

  13 |     ARG GROUPID

  14 | >>> RUN adduser --uid $USERID --gid $GROUPID --disabled-password --gecos "" nonroot

  15 |     RUN echo 'nonroot ALL=(ALL) NOPASSWD: ALL' >> /etc/sudoers

  16 |     USER nonroot

--------------------

target display: failed to solve: process "/bin/sh -c adduser --uid $USERID --gid $GROUPID --disabled-password --gecos \"\" nonroot" did not complete successfully: exit code: 1

```

I am on fedora linux with

```bash
$ docker --version
Docker version 28.5.1, build 2.fc42
```

```bash
$ docker compose version
Docker Compose version 2.40.0
```
