### 2025-10-31 #9218 @magland
@skanwal oh sorry... I ran into another issue in trying to get started and submitted another ticket
https://github.com/TACC/tissue-data-explorer/issues/19
