### 2025-11-03 TACC/tissue-data-explorer#19 @james-labyer
Hi there, I have updated the Dockerfile to fix this issue. Please download the most recent version of `Dockerfile-all` and try rebuilding the app. Let me know if you run into any more issues.
