### 2025-11-11 a2ea6d7
Updated README for clarity
