+++ README.md
@@ -3,42 +3,21 @@
 This project contains the code necessary to create a website that showcases scientific data collected on a tissue sample. Types of data handled by this example include multi-channel image stacks, volumetric map data, and 3D models. This repository includes synthetic data that can be used to stand up a demo website, as well as a configuration portal that can be used to upload project data to the site.
 
 ## Prerequisites
-Docker/ Docker Compose
 
-## Getting Started with Development
+- Docker Desktop (version 4.43 or higher) or Docker Engine (version 28.3 or higher)
+- Docker Compose (version 2.38 or higher)
+
+## Preparing the production image
+
 1. Clone this repo
 
    ```
    git clone https://github.com/TACC/tissue-data-explorer.git
    ```
 
-2. Add a `.env` file to the `config_portal` folder. See `.env.example` for format.
+2. Add a `.env` file to the `config_portal` folder. See the section `Log in credentials for configuration site` for more information.
 
-3. Use the script `./build.sh` to build a development environment. If this is your first time building a development environment, you will need to create a new volume for use with the environment by providing the `-n` option to the program. The default behavior is to populate your new volume with some test data: 
-   ```
-   ./build.sh -n
-   ```
-
-   Or, if you would prefer to use a minimum dataset, you can specify that by providing the value `min` to the `-d` option. 
-   ```
-   ./build.sh -n -d min
-   ```
-
-   You can also specify a custom volume name using the `-v` option. 
-
-   Running the script builds the apps and starts the display app at `localhost:8050` and the config app at `localhost:8040`.
-
-
-## Running tests locally
-The script `run_tests.sh` in the root project folder creates docker containers for the display and configuration apps, fills them with test data, runs the tests for the display app and config app, and then deletes the test containers and test volume.
-
-```
-./run_tests.sh
-```
-
-## Preparing the production image
-
-1. On a machine with this project's GitHub repo, build the images. The build script as currently configured will create the production build for the `linux/amd64` and `linux/arm64` platforms. If neither of those platforms meet your needs, then you will need to update the platforms specified in `docker-compose.yaml` to fit the platform of your production server. See the Docker Compose [documentation](https://docs.docker.com/reference/compose-file/build/#platforms) for more information.
+3. Build the images. The build script as currently configured will create the production build for the `linux/amd64` and `linux/arm64` platforms. If neither of those platforms meet your needs, then you will need to update the platforms specified in `docker-compose.yaml` to fit the platform of your production server. See the Docker Compose [documentation](https://docs.docker.com/reference/compose-file/build/#platforms) for more information.
 
    Providing the value `prod` to the `-d` option of the build script will trigger a production build. You can specify a custom volume using the `-v` option as well.
 
@@ -46,7 +25,7 @@ The script `run_tests.sh` in the root project folder creates docker containers f
    ./build.sh -e prod
    ```
 
-2. Replace the variables in the code snippet below with your username and the appropriate version tag and run the commands to publish the images to Docker Hub
+4. Replace the variables in the code snippet below with your username and the appropriate version tag and run the commands to publish the images to Docker Hub
 
    ```
    docker tag tde-prod-display {username}/tissue-data-explorer-display:{tag}
@@ -55,18 +34,19 @@ The script `run_tests.sh` in the root project folder creates docker containers f
    docker push {username}/tissue-data-explorer-config:{tag}
    ```
 
-3. On the production server, pull the newly published images from Docker Hub
+5. On the production server, pull the newly published images from Docker Hub
 
    ```
    docker pull {username}/tissue-data-explorer-display:{tag}
    docker pull {username}/tissue-data-explorer-config:{tag}
    ```
 
-4. Copy the `docker-compose-prod.yaml` file and your `.env` file for the production configuration app onto the production server. You will need to make the following changes to `docker-compose-prod.yaml`:
[…]
