### 2025-12-08 TACC/tissue-data-explorer#19 @magland
Hi @james-labyer sorry for the delay, I somehow missed your reply.

With the new version, I ran

./build.sh -n

and that started up two containers: tde-dev-config and tde-dev-display

but when I open in chrome http://localhost:8040 and http://localhost:8050 as in the instructions I get "site cannot be reached".

I'm on fedora linux laptop with

```bash
$ docker --version
Docker version 29.0.4, build 1.fc42

$ docker compose version
Docker Compose version 2.40.3
```
