### 2025-12-08 TACC/tissue-data-explorer#19 @james-labyer
Did you create the `.env` file in the config_portal directory? You can copy and rename the `.env.example` file in that directory for testing purposes. If your `.env` file is in order, please run `docker logs <container>` for each container and share the output with me.
