### 2025-12-09 TACC/tissue-data-explorer#19 @magland
Here's what I get

```
$ docker logs tde-dev-config-1
python: can't open file '/home/nonroot/app/app.py': [Errno 13] Permission denied

$ docker logs tde-dev-display-1
python: can't open file '/home/nonroot/app/app.py': [Errno 13] Permission denied
```
