### 2025-12-15 TACC/tissue-data-explorer#19 @james-labyer
Hi there. I am running the app on a VM with Fedora 43 and it builds and runs with no issues. I am wondering what is different between your environment and mine. Can you please check your Fedora version using `cat /etc/fedora-release`? Also, I'm wondering if you are running as a user with limited permissions. Are you running this in a development environment where you have a lot of control over the environment, or are you running it in a shared system?
