### 2025-12-15 TACC/tissue-data-explorer#19 @magland
Hi @james-labyer

For fedora release I get
Fedora release 42 (Adams)

But I am using an immutable operating system  (silverblue) where the system os is mounted as read only. Maybe that's causing the permissions issue?
