### 2025-12-16 TACC/tissue-data-explorer#19 @james-labyer
We were able to follow the project installation instructions on a VM running silverblue after following these docker installation instructions, which were found in [this thread](https://discussion.fedoraproject.org/t/installing-docker-on-silverblue/119610/5):

```
1. curl -O https://download.docker.com/linux/fedora/docker-ce.repo
2. sudo install -o 0 -g 0 -m644 docker-ce.repo /etc/yum.repos.d/docker-ce.repo
3. sudo systemctl reboot
4. sudo rpm-ostree install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
5. sudo systemctl reboot
6. sudo systemctl enable docker
7. sudo systemctl reboot
8. grep -E "^docker:" /usr/lib/group | sudo tee -a /etc/group && sudo usermod -aG "docker" "$USER"
9. sudo systemctl reboot
10. Follow tissue-data-explorer dev install instructions
```

If updating your docker installation doesn't resolve the issue, would you be able to use a VM to test the app in another operating system? We didn't develop or test the app with an immutable operating system so it's hard to know how much troubleshooting will be required if this suggestion doesn't work.

We have tested the app successfully with the following operating systems:
- macOS sonoma 14.7
- Fedora Workstation release 43
- Ubuntu 24.04.1 LTS

Could you please try testing with a VM running a similar operating system? Thank you for your continued time on this review.
