### 2025-12-18 TACC/tissue-data-explorer#19 @magland
@james-labyer thanks for those detailed instructions. Unfortunately after working through those steps, I still get the same error. One factor could be that I already had docker installed... not sure if that played a role.

I understand I'm on an unusual system, but I think the spirit of reviewing for JOSS is that I wouldn't need to do a lot of troubleshooting. Perhaps a different reviewer could be chosen with a more standard environment? Or I'm willing to go the VM route, but I would need detailed instructions on how to do that. I have never used a VM.

I was hoping I could just follow a set of instructions and evaluate on the basis of that. Didn't realize my OS setup would be a problem.
