### 2026-01-06 TACC/tissue-data-explorer#19 @james-labyer
Thanks for the response- we didn't anticipate a setup like yours, so we didn't expect the OS setup to be a barrier either. Thanks for bearing with us.

If we set up a VM for you, would you be willing to use that for testing? You would just need to send us your public key so we can give you access to the VM. Here are [instructions](https://docs.github.com/en/authentication/connecting-to-github-with-ssh/generating-a-new-ssh-key-and-adding-it-to-the-ssh-agent?platform=linux#generating-a-new-ssh-key) on how to generate a public key.
