### 2026-01-13 TACC/tissue-data-explorer#19 @magland
Hi @james-labyer

I have generated a public key...

ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIDR8fkfM3Wjv4p5VXbZ6eiE8RinOXtkswwQzE4cD71fC jmagland@flatironinstitute.org
