### 2026-01-16 TACC/tissue-data-explorer#19 @james-labyer
Hi there, we have set up a VM using Ubuntu 24.04 for you. You can connect to the VM using the command `ssh exouser@129.114.39.53`. Once you have connected, you can set up the Tissue Data Explorer project using the application [documentation](https://github.com/TACC/tissue-data-explorer/blob/main/README.md#getting-started-with-development). Once the app is up and running, you should be able to access it in the browser on your machine by navigating to `http://129.114.39.53:[configured port]`. Let us know if you have any questions.
