### 2026-01-27 #9218 @magland
With the help of the authors kindly providing me with a virtual machine in the cloud for testing, I was able to confirm the documented installation process and verify the functionality of the software. I have completed the checklist.
