### 2026-02-09 cf13ffc
Update paper.bib
