+++ paper/paper.bib
@@ -36,7 +36,8 @@
 }
 
 @misc{Stack:2024,
-    title = {Stack Overflow Developer Survey 2024},
+    title = {Stack Overflow Developer Survey},
+    year = {2024},
     url={https://survey.stackoverflow.co/2024/},
     journal = {Stack Exchange Inc}
 }
