### 2026-02-09 b8adb3f
Update paper.md
