### 2026-02-09 d3b4794
Create CITATION.cff
