### 2026-02-09 #9218 @james-labyer
We have completed all of the author review tasks:

The software version is v.1.0.0 and the Zenodo DOI is https://doi.org/10.5281/zenodo.18567723 .

Thanks again for all of your help!
