### 2026-02-18 #9218 @james-labyer
Hi there, the suggested DOI 10.1007/978-1-4842-3012-1_9 under "Missing DOIs" refers to an [instruction manual](https://link.springer.com/chapter/10.1007/978-1-4842-3012-1_9) written about Docker Compose by a third party, not the [Docker Compose software](https://github.com/docker/compose) itself, which is what the citation should refer to. The Docker Compose software does not have a DOI associated with it.
