+++ paper/paper.md
@@ -45,7 +45,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Tissue Data Explorer is a Python code repository that can be used to deploy a website for displaying 2D and 3D biomolecular datasets that have been collected from tissue samples taken from organs [@Labyer:2025]. The Tissue Data Explorer includes interfaces for loading and displaying image stacks, spatial maps of measurements taken from different areas within the tissue sample, 3D models of the tissue sample, and links to custom reports. Life sciences researchers can deploy an instance of the website as is, or they can fork the repository and modify the interfaces to accommodate additional data types. To enable easier customization, the website is written using the Plotly Dash visualization framework [@Parmer:2024]. Dash allows users to build visuals using Python, which is likely to be easier for many biologists because Python knowledge is common among life sciences researchers [@Mariano:2020] [@Boudreau:2021]. This is in contrast with other common web languages like JavaScript, which is popular among software developers [@Stack:2024] but has not achieved the same level of adoption with life sciences researchers [@Sorani:2012].
+The Tissue Data Explorer is a Python code repository that can be used to deploy a website for displaying 2D and 3D biomolecular datasets that have been collected from tissue samples taken from organs. The Tissue Data Explorer includes interfaces for loading and displaying image stacks, spatial maps of measurements taken from different areas within the tissue sample, 3D models of the tissue sample, and links to custom reports. Life sciences researchers can deploy an instance of the website as is, or they can fork the repository and modify the interfaces to accommodate additional data types. To enable easier customization, the website is written using the Plotly Dash visualization framework [@Parmer:2024]. Dash allows users to build visuals using Python, which is likely to be easier for many biologists because Python knowledge is common among life sciences researchers [@Mariano:2020] [@Boudreau:2021]. This is in contrast with other common web languages like JavaScript, which is popular among software developers [@Stack:2024] but has not achieved the same level of adoption with life sciences researchers [@Sorani:2012].
 
 # Statement of Need
 
