### 2026-03-04 144846c
Update paper.md
Fixed figure label duplicate
