### 2026-03-04 #9218 @jamescarson3
@skanwal I think I fixed the issue with the duplicate figure label in paper.md
