### 2026-03-23 #9218 @Kevin-Mattheus-Moerman
@jamescarson3 As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which require your attention:

👉 Please edit the title over at the archive to match: `Tissue Data Explorer: a website template for presenting tissue sample research findings`

👉 Please add the country information to your affiliations. I.e. add `United States of America` (spelled out like this) to each.
-> yes, it looks like we have a wrongly detected missing DOI. The authors intend to cite a GitHub repository. The identified DOI is for a book chapter with the same name.
- [x] Copyedit pass/checked for typos
